LayoutTests/imported/w3c/web-platform-tests/custom-elements/registries/ShadowRoot-init-customElementRegistry-expected.txt


PASS A newly attached disconnected ShadowRoot should use the global registry by default
PASS A newly attached connected ShadowRoot should use the global registry by default
PASS A newly attached disconnected ShadowRoot should use the scoped registry if explicitly specified in attachShadow
PASS A newly attached connected ShadowRoot should use the scoped registry if explicitly specified in attachShadow
PASS attachShadow() should use the global registry when customElementRegistry is null
FAIL attachShadow() should use the shadow host's registry when customElementRegistry is null assert_equals: expected object "[object CustomElementRegistry]" but got object "[object CustomElementRegistry]"
PASS attachShadow() should use the null registry when the shadow host uses null registry and customElementRegistry is null
